// src/app/models/types.ts
export enum QuestionType {
  TEXT = "TEXT",
  MAP_FEATURE = "MAP_FEATURE",
  MAP_FEATURE_FREE = "MAP_FEATURE_FREE",
  NAV_INSTRUCTION = "NAV_INSTRUCTION",
  MAP_TARGET = "MAP_TARGET",
  MAP_DIRECTION = "MAP_DIRECTION",
  MAP_DIRECTION_MARKER = "MAP_DIRECTION_MARKER",
  MAP_FEATURE_PHOTO = "MAP_FEATURE_PHOTO",
  MAP_DIRECTION_PHOTO = "MAP_DIRECTION_PHOTO",
  INFO = "INFO",
  PHOTO = "PHOTO",
  NAV_INSTRUCTION_PHOTO = "NAV_INSTRUCTION_PHOTO",
}
export enum AnswerType {
  EXPLORATION = "EXPLORATION",
  POSITION = "POSITION",
  MAP_POINT = "MAP_POINT",
  MULTIPLE_CHOICE = "MULTIPLE_CHOICE",
  PHOTO = "PHOTO",
  DIRECTION = "DIRECTION",
  MAP_DIRECTION = "MAP_DIRECTION",
  INFO = "INFO",
  MULTIPLE_CHOICE_TEXT = "MULTIPLE_CHOICE_TEXT",
  TEXT = "TEXT",
  NUMBER = "NUMBER",
  DRAW = "DRAW",
}

export enum TaskMode {
  NAV_ARROW = "NAV_ARROW",
  DIRECTION_ARROW = "DIRECTION_ARROW",
  NO_FEATURE = "NO_FEATURE",
}